+++ src/main/java/org/betonquest/betonquest/variables/TagVariable.java
@@ -1,0 +1,72 @@
+package org.betonquest.betonquest.variables;
+
+import org.betonquest.betonquest.BetonQuest;
+import org.betonquest.betonquest.Instruction;
+import org.betonquest.betonquest.api.Variable;
+import org.betonquest.betonquest.api.config.QuestPackage;
+import org.betonquest.betonquest.config.Config;
+import org.betonquest.betonquest.exceptions.InstructionParseException;
+
+import java.util.List;
+
+/**
+ * Exposes the presence of tags as a variable.
+ * Originally implemented for use with the PAPI integration.
+ */
+public class TagVariable extends Variable {
+
+    /**
+     * The tag to check for.
+     */
+    private final String tagName;
+    /**
+     * The quest package to check for the tag.
+     */
+    private final QuestPackage questPackage;
+
+    /**
+     * Whether to return true/false or the configured messages.
+     */
+    private final boolean papiMode;
+
+    /**
+     * Constructs a new GlobalTagVariable.
+     *
+     * @param instruction the instruction to parse
+     * @throws InstructionParseException if the instruction is malformed
+     */
+    public TagVariable(final Instruction instruction) throws InstructionParseException {
+        super(instruction);
+        tagName = instruction.next();
+        questPackage = instruction.getPackage();
+        papiMode = instruction.hasArgument("papiMode");
+    }
+
+    /**
+     * Returns the value of the variable.
+     *
+     * @param playerID ID of the player
+     * @return the value of the variable
+     */
+    @Override
+    public String getValue(final String playerID) {
+        final List<String> tags = BetonQuest.getInstance().getPlayerData(playerID).getTags();
+        return getValue(tags);
+    }
+
+    /**
+     * Finds out whether the player has the tag and returns this as a true/false string.
+     * If papiMode is enabled, it returns the configured messages instead.
+     *
+     * @param tags the tags to check
+     * @return whether the player has the variable
+     */
+    protected String getValue(final List<String> tags) {
+        final String lang = Config.getLanguage();
+
+        if (tags.contains(questPackage.getPackagePath() + "." + tagName)) {
+            return papiMode ? Config.getMessage(lang, "condition_variable_met") : "true";
+        }
+        return papiMode ? Config.getMessage(lang, "condition_variable_not_met") : "false";
+    }
+}